[Carlo][chore] renamed the kanban board
